trie: move orderBatch function to util file / delete batch.ts
